Fixes issue #3349 - Bump code coverage
